% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leapyear.R
\name{is_leapyear}
\alias{is_leapyear}
\title{Check if a given year is a leap year}
\usage{
is_leapyear(year)
}
\arguments{
\item{year}{A numeric or numeric-string scalar input representing a year.}
}
\value{
TRUE if it is a leap year, FALSE otherwise.
}
\description{
This function determines whether the provided year is a leap year
based on the rules of the Gregorian calendar.
}
\examples{
is_leapyear(2000) # TRUE
is_leapyear("1992") # TRUE
is_leapyear("1900") # FALSE

}
